+++ release_notes.md
@@ -1,6 +1,7 @@
 ## New Features
 
 ## Bug fixes
+- Fixed handling of function timeouts inside entity and activity functions and added tests
 
 ## Breaking Changes